services/trade_engine/requirements.txt
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.28.0
websockets==14.1
alpaca-py==0.33.1
